playback: split support types by responsibility
