+++ Sources/SpeakSwiftly/Playback/PlaybackThresholdController.swift
@@ -1,135 +1,4 @@
-@preconcurrency import AppKit
-@preconcurrency import AVFoundation
-import CoreAudio
 import Foundation
-import TextForSpeech
-
-// MARK: - AudioPlaybackConfiguration
-
-enum AudioPlaybackConfiguration {
-    static let minimumDrainTimeout: Duration = .seconds(5)
-    static let drainTimeoutPaddingMS = 3000
-    static let drainProgressCheckIntervalMS = 500
-    static let drainProgressStallTimeoutMS = 8000
-    static let environmentInstabilityWindowMS = 8000
-    static let recoveryStabilizationDelayMS = 900
-    static let recoveryMaximumAttempts = 3
-    static let lowQueueThresholdMS = 100
-    static let channels: AVAudioChannelCount = 1
-    static let interJobBoopDurationMS = 90
-    static let interJobBoopFrequencyHz = 1176.0
-    static let interJobBoopAmplitude: Float = 0.14
-    static let interJobBoopFadeMS = 10
-    static let interJobBoopTimeout: Duration = .seconds(2)
-
-    static func drainTimeout(forQueuedAudioMS queuedAudioMS: Int) -> Duration {
-        let paddedQueuedAudioMS = queuedAudioMS + drainTimeoutPaddingMS
-        return max(minimumDrainTimeout, .milliseconds(paddedQueuedAudioMS))
-    }
-}
-
-// MARK: - Playback Device Inspection
-
-func currentDefaultAudioPlaybackDeviceDescription() -> String? {
-    var deviceID = AudioObjectID(kAudioObjectUnknown)
-    var dataSize = UInt32(MemoryLayout<AudioObjectID>.size)
-    var deviceAddress = AudioObjectPropertyAddress(
-        mSelector: kAudioHardwarePropertyDefaultOutputDevice,
-        mScope: kAudioObjectPropertyScopeGlobal,
-        mElement: kAudioObjectPropertyElementMain,
-    )
-    let deviceStatus = AudioObjectGetPropertyData(
-        AudioObjectID(kAudioObjectSystemObject),
-        &deviceAddress,
-        0,
-        nil,
-        &dataSize,
-        &deviceID,
-    )
-
-    guard deviceStatus == noErr, deviceID != AudioObjectID(kAudioObjectUnknown) else {
-        return nil
-    }
-
-    var deviceName: CFString = "" as CFString
-    var nameSize = UInt32(MemoryLayout<CFString>.stride)
-    var nameAddress = AudioObjectPropertyAddress(
-        mSelector: kAudioDevicePropertyDeviceNameCFString,
-        mScope: kAudioObjectPropertyScopeGlobal,
-        mElement: kAudioObjectPropertyElementMain,
-    )
-    let nameStatus = withUnsafeMutablePointer(to: &deviceName) { pointer in
-        AudioObjectGetPropertyData(
-            deviceID,
-            &nameAddress,
-            0,
-            nil,
-            &nameSize,
-            UnsafeMutableRawPointer(pointer),
-        )
-    }
-
-    if nameStatus == noErr {
-        let name = "\(deviceName)"
-        if !name.isEmpty {
-            return "\(name) [\(deviceID)]"
-        }
-    }
-
-    return "AudioObjectID \(deviceID)"
-}
-
-// MARK: - PlaybackEvent
-
-enum PlaybackEvent {
-    case firstChunk
-    case prerollReady(startupBufferedAudioMS: Int, thresholds: PlaybackAdaptiveThresholds)
-    case queueDepthLow(queuedAudioMS: Int)
-    case rebufferStarted(queuedAudioMS: Int, thresholds: PlaybackAdaptiveThresholds)
-    case rebufferResumed(bufferedAudioMS: Int, thresholds: PlaybackAdaptiveThresholds)
-    case chunkGapWarning(gapMS: Int, chunkIndex: Int)
-    case scheduleGapWarning(gapMS: Int, bufferIndex: Int, queuedAudioMS: Int)
-    case rebufferThrashWarning(rebufferEventCount: Int, windowMS: Int)
-    case outputDeviceChanged(previousDevice: String?, currentDevice: String?)
-    case engineConfigurationChanged(engineIsRunning: Bool)
-    case bufferShapeSummary(
-        maxBoundaryDiscontinuity: Double,
-        maxLeadingAbsAmplitude: Double,
-        maxTrailingAbsAmplitude: Double,
-        fadeInChunkCount: Int,
-    )
-    case trace(PlaybackTraceEvent)
-    case starved
-}
-
-// MARK: - PlaybackEnvironmentEvent
-
-enum PlaybackEnvironmentEvent {
-    case outputDeviceObserved(currentDevice: String?)
-    case outputDeviceChanged(previousDevice: String?, currentDevice: String?)
-    case engineConfigurationChanged(engineIsRunning: Bool)
-    case systemSleepStateChanged(isSleeping: Bool)
-    case screenSleepStateChanged(isSleeping: Bool)
-    case sessionActivityChanged(isActive: Bool)
-    case recoveryStateChanged(reason: String, stage: String, attempt: Int?, currentDevice: String?)
-    case interJobBoopPlayed(durationMS: Int, frequencyHz: Double, sampleRate: Double)
-}
-
-// MARK: - PlaybackTraceEvent
-
-struct PlaybackTraceEvent {
-    let name: String
-    let chunkIndex: Int?
-    let bufferIndex: Int?
-    let sampleCount: Int?
-    let durationMS: Int?
-    let queuedAudioBeforeMS: Int?
-    let queuedAudioAfterMS: Int?
-    let gapMS: Int?
-    let isRebuffering: Bool?
-    let fadeInApplied: Bool?
-}
-
 // MARK: - PlaybackComplexityClass
 
 enum PlaybackComplexityClass: String {
@@ -745,209 +614,4 @@ struct PlaybackThresholdController {
 
         return values.reduce(0, +) / values.count
     }
-}
-
-// MARK: - PlaybackMetricsConfiguration
-
-enum PlaybackMetricsConfiguration {
-    static let rebufferThrashWarningCount = 3
-    static let rebufferThrashWindowMS = 2000
-}
-
-// MARK: - PlaybackSummary
-
-struct PlaybackSummary {
-    let thresholds: PlaybackAdaptiveThresholds
-    let chunkCount: Int
-    let sampleCount: Int
-    let startupBufferedAudioMS: Int?
-    let timeToFirstChunkMS: Int?
-    let timeToPrerollReadyMS: Int?
-    let timeFromPrerollReadyToDrainMS: Int?
-    let minQueuedAudioMS: Int?
-    let maxQueuedAudioMS: Int?
-    let avgQueuedAudioMS: Int?
-    let queueDepthSampleCount: Int
-    let rebufferEventCount: Int
-    let rebufferTotalDurationMS: Int
-    let longestRebufferDurationMS: Int
-    let starvationEventCount: Int
-    let scheduleCallbackCount: Int
-    let playedBackCallbackCount: Int
-    let maxInterChunkGapMS: Int?
-    let avgInterChunkGapMS: Int?
-    let maxScheduleGapMS: Int?
-    let avgScheduleGapMS: Int?
-    let maxBoundaryDiscontinuity: Double?
-    let maxLeadingAbsAmplitude: Double?
-    let maxTrailingAbsAmplitude: Double?
-    let fadeInChunkCount: Int
-}
-
-// MARK: - RuntimeMemorySnapshot
-
-struct RuntimeMemorySnapshot {
-    let processResidentBytes: Int?
-    let processPhysFootprintBytes: Int?
-    let processUserCPUTimeNS: Int?
-    let processSystemCPUTimeNS: Int?
-    let mlxActiveMemoryBytes: Int?
-    let mlxCacheMemoryBytes: Int?
-    let mlxPeakMemoryBytes: Int?
-    let mlxCacheLimitBytes: Int?
-    let mlxMemoryLimitBytes: Int?
-}
-
-// MARK: - AudioPlaybackRequestState
-
-@MainActor
-final class AudioPlaybackRequestState {
-    struct QueuedBuffer {
-        let pcmBuffer: AVAudioPCMBuffer
-        let frameCount: Int
-        let firstSample: Float
-        let lastSample: Float
-        let fadeInApplied: Bool
-        let chunkIndex: Int
-        var bufferIndex: Int?
-        var engineGeneration: Int?
-    }
-
-    let requestID: UInt64
-    var thresholdsController: PlaybackThresholdController
-    var generationFinished = false
-    var isRebuffering = false
-    var queuedBuffers = [QueuedBuffer]()
-    var queuedSampleCount = 0
-    var nextBufferIndex = 0
-    var engineGeneration = 0
-    var minQueuedAudioMS: Int?
-    var maxQueuedAudioMS: Int?
-    var queueDepthTotalMS = 0
-    var queueDepthSampleCount = 0
-    var rebufferEventCount = 0
-    var rebufferStartedAt: Date?
-    var rebufferTotalDurationMS = 0
-    var longestRebufferDurationMS = 0
-    var recentRebufferStartTimes = [Date]()
-    var emittedRebufferThrashWarning = false
-    var starvationEventCount = 0
-    var emittedLowQueueWarning = false
-    var scheduleCallbackCount = 0
-    var playedBackCallbackCount = 0
-    var lastTrailingSample: Float?
-    var maxBoundaryDiscontinuity: Double?
-    var maxLeadingAbsAmplitude: Double?
-    var maxTrailingAbsAmplitude: Double?
-    var fadeInChunkCount = 0
-    var drainContinuation: CheckedContinuation<Void, Error>?
-
-    init(requestID: UInt64, text: String, tuningProfile: PlaybackTuningProfile) {
-        self.requestID = requestID
-        thresholdsController = PlaybackThresholdController(text: text, tuningProfile: tuningProfile)
-    }
-
-    func queuedAudioMS(sampleRate: Double) -> Int {
-        Int((Double(max(queuedSampleCount, 0)) / sampleRate * 1000).rounded())
-    }
-
-    func recordQueuedAudioDepth(sampleRate: Double) {
-        let currentQueuedAudioMS = queuedAudioMS(sampleRate: sampleRate)
-        minQueuedAudioMS = min(minQueuedAudioMS ?? currentQueuedAudioMS, currentQueuedAudioMS)
-        maxQueuedAudioMS = max(maxQueuedAudioMS ?? currentQueuedAudioMS, currentQueuedAudioMS)
-        queueDepthTotalMS += currentQueuedAudioMS
-        queueDepthSampleCount += 1
-    }
-
-    func installDrainContinuation(
-        _ continuation: CheckedContinuation<Void, Error>,
-        sampleRate: Double,
-    ) {
-        drainContinuation = continuation
-        if queuedAudioMS(sampleRate: sampleRate) == 0 {
-            resumeDrainContinuation()
-        }
-    }
-
-    func resumeDrainContinuation() {
-        guard let drainContinuation else { return }
-
-        self.drainContinuation = nil
-        drainContinuation.resume()
-    }
-
-    func resumeDrainContinuation(throwing error: any Error) {
-        guard let drainContinuation else { return }
-
-        self.drainContinuation = nil
-        drainContinuation.resume(throwing: error)
-    }
-
-    func enqueueBuffer(
-        _ pcmBuffer: AVAudioPCMBuffer,
-        frameCount: Int,
-        firstSample: Float,
-        lastSample: Float,
-        fadeInApplied: Bool,
-        chunkIndex: Int,
-    ) {
-        queuedBuffers.append(
-            QueuedBuffer(
-                pcmBuffer: pcmBuffer,
-                frameCount: frameCount,
-                firstSample: firstSample,
-                lastSample: lastSample,
-                fadeInApplied: fadeInApplied,
-                chunkIndex: chunkIndex,
-                bufferIndex: nil,
-                engineGeneration: nil,
-            ),
-        )
-        queuedSampleCount += frameCount
-    }
-
-    func reserveQueuedBufferIndicesForCurrentGeneration() -> [QueuedBuffer] {
-        guard !queuedBuffers.isEmpty else { return [] }
-
-        var reserved = [QueuedBuffer]()
-        for index in queuedBuffers.indices where queuedBuffers[index].bufferIndex == nil {
-            let bufferIndex = nextBufferIndex + 1
-            nextBufferIndex = bufferIndex
-            queuedBuffers[index].bufferIndex = bufferIndex
-            queuedBuffers[index].engineGeneration = engineGeneration
-            reserved.append(queuedBuffers[index])
-        }
-        return reserved
-    }
-
-    func markQueuedBuffersForReschedule() {
-        engineGeneration += 1
-        for index in queuedBuffers.indices {
-            queuedBuffers[index].bufferIndex = nil
-            queuedBuffers[index].engineGeneration = nil
-        }
-    }
-
-    func completeQueuedBuffer(
-        bufferIndex: Int,
-        engineGeneration: Int,
-    ) -> QueuedBuffer? {
-        guard let queueIndex = queuedBuffers.firstIndex(where: {
-            $0.bufferIndex == bufferIndex && $0.engineGeneration == engineGeneration
-        }) else {
-            return nil
-        }
-
-        let completedBuffer = queuedBuffers.remove(at: queueIndex)
-        queuedSampleCount = max(0, queuedSampleCount - completedBuffer.frameCount)
-        return completedBuffer
-    }
-}
-
-// MARK: - AudioPlaybackRecoveryReason
-
-enum AudioPlaybackRecoveryReason: String {
-    case systemWake = "system_wake"
-    case outputDeviceChange = "output_device_change"
-    case engineConfigurationChange = "engine_configuration_change"
 }
